Add check input
